+++ test/integration_test.exs
@@ -1,6 +1,9 @@
 defmodule KazanIntegrationTest do
   use ExUnit.Case, async: true
 
+  alias Kazan.Apis.Core.V1, as: CoreV1
+  alias Kazan.Apis.Extensions.V1beta1, as: ExtensionsV1beta1
+  alias Kazan.Apis.Rbacauthorization.V1beta1, as: RbacauthorizationV1beta1
   alias Kazan.Models.Core.V1
   alias Kazan.Models.Apimachinery.Meta.V1.{
     ObjectMeta,
@@ -21,7 +24,7 @@ defmodule KazanIntegrationTest do
 
   test "can list namespaces on an actual server", %{server: server} do
     namespace_list =
-      Kazan.Apis.CoreV1.list_namespace!
+      CoreV1.list_namespace!
       |> Kazan.Client.run!(server: server)
 
     # Check that there's a default namespace.
@@ -31,18 +34,18 @@ defmodule KazanIntegrationTest do
   end
 
   test "can list pods on an actual server", %{server: server} do
-    Kazan.Apis.CoreV1.list_namespaced_pod!(@namespace)
+    CoreV1.list_namespaced_pod!(@namespace)
     |> Kazan.Client.run!(server: server)
   end
 
   test "can list deployments on an actual server", %{server: server} do
-    Kazan.Apis.ExtensionsV1beta1.list_namespaced_deployment!(@namespace)
+    ExtensionsV1beta1.list_namespaced_deployment!(@namespace)
     |> Kazan.Client.run!(server: server)
   end
 
   test "can create, patch and delete a pod", %{server: server} do
     created_pod =
-      Kazan.Apis.CoreV1.create_namespaced_pod!(
+      CoreV1.create_namespaced_pod!(
         %V1.Pod{
           metadata: %ObjectMeta{name: "kazan-test"},
           spec: %V1.PodSpec{
@@ -60,12 +63,12 @@ defmodule KazanIntegrationTest do
       |> Kazan.Client.run!(server: server)
 
     read_pod =
-      Kazan.Apis.CoreV1.read_namespaced_pod!(@namespace, "kazan-test")
+      CoreV1.read_namespaced_pod!(@namespace, "kazan-test")
       |> Kazan.Client.run!(server: server)
 
     assert read_pod.spec == %{created_pod.spec | node_name: read_pod.spec.node_name}
 
-    patched_pod = Kazan.Apis.CoreV1.patch_namespaced_pod!(
+    patched_pod = CoreV1.patch_namespaced_pod!(
       %V1.Pod{
         metadata: %ObjectMeta{name: "kazan-test"},
         spec: %V1.PodSpec{
@@ -79,12 +82,12 @@ defmodule KazanIntegrationTest do
     assert patched_pod.spec == %{read_pod.spec | active_deadline_seconds: 5}
 
     read_pod =
-      Kazan.Apis.CoreV1.read_namespaced_pod!(@namespace, "kazan-test")
+      CoreV1.read_namespaced_pod!(@namespace, "kazan-test")
       |> Kazan.Client.run!(server: server)
 
     assert read_pod == patched_pod
 
-    Kazan.Apis.CoreV1.delete_namespaced_pod!(
+    CoreV1.delete_namespaced_pod!(
       %DeleteOptions{}, @namespace, "kazan-test"
     )
     |> Kazan.Client.run!(server: server)
@@ -92,7 +95,7 @@ defmodule KazanIntegrationTest do
 
   test "RBAC Authorization V1 Beta 1 API", %{server: server} do
     cluster_roles =
-      Kazan.Apis.RbacAuthorizationV1beta1.list_cluster_role!()
+      RbacauthorizationV1beta1.list_cluster_role!()
       |> Kazan.Client.run!(server: server)
 
     assert cluster_roles.kind == "ClusterRoleList"